fix : 폭발열매 맞았을 때 맞는 애니메이션 적용

+++ Roguelike/Assets/2.Scripts/Enemy/DamageText.cs
@@ -11,7 +11,7 @@ public class DamageText : MonoBehaviour
     
     private TextMeshPro text;
     private Color alpha;
-    public int damage; // 임시 코드 변경할 예정
+    public float damage; // 임시 코드 변경할 예정
     
     void Start()
     {
